Typing `ApiRoutes`, fixing fallout.
